Issue #9: Move the modifiers from the home row to the row above it
This is not as nice but I realized the home row was already taken by the Vim
movement keys and I really want to be able to hold them and have the cursor
going down the screen. Having the modifiers on home row forced me to tap the
keys instead otherwise they were interpreted as modifiers.

Still not used to type like this though...